Update README.mdlimtedSUPPLYnft
The FoundryCourseNft contract is a robust implementation of a token minting mechanism for a specific type of non-fungible token (NFT). This contract is written in Solidity, targeting the Ethereum blockchain.

The primary functionality of this smart contract includes:

A minting function mintNft() that allows the creation of a new NFT. The function requires an Ethereum address and a Twitter handle, which is then mapped to the NFT. The function is payable, implying it can be used to require a payment to mint an NFT. It returns the token ID of the newly minted token.

A maximum supply limit that prevents the minting of more than 10,000 NFTs. This is enforced through a check in the mintNft() function. This implementation guarantees a hard cap on the total NFTs in circulation, adding a rarity aspect.

A function addChallenge() that allows adding a new challenge contract address to an array challenges. This functionality could be used to add supplementary contracts that interact with the NFT contract in terms of providing additional challenges or achievements for the NFT holders.

The contract also includes a fallback function and a receive function. These functions are triggered when the contract receives ether but no function is called or if the called function does not exist.

The state of the contract, such as the twitter handles of the NFT owners, balances of each address, and the challenges added, are publicly accessible.